+++ src/main/java/com/ruse/world/content/combat/magic/CombatNormalSpell.java
@@ -11,6 +11,6 @@ import com.ruse.world.entity.impl.Character;
 public abstract class CombatNormalSpell extends CombatSpell {
 
 	@Override
-	public void finishCast(Character cast, Character castOn, boolean accurate, int damage) {
+	public void finishCast(Character cast, Character castOn, boolean accurate, long damage) {
 	}
 }
